requirements.txt
loguru
exa_py
tiktoken
python-dotenv
streamlit